Dont run workflow jobs twice if master is pushed
The idea is to push on feature branches
Tests happen there
Then merge it into the master, it should be impossible to slip a mistake in since the feature is tested before merging.

BUT even if a mistake slips in, its still blocked by the autocheck of octodns sync command.